refactor: Move getWallPaperUrlByRace to utils.js
